Tener tiempos por vuelta de cada jugador: listo

+++ PolePosition2.0.2/Assets/_Scripts/_Player/PlayerInfo.cs
@@ -10,8 +10,8 @@ public class PlayerInfo : MonoBehaviour
     public int CarColorID;
     public int CurrentPosition;
     public int CurrentLap;
-    public int CurrentLapTime;
-    public int CurrentRaceTime;
+    public float CurrentLapTime;
+    public float CurrentRaceTime;
     public int LastCheckpoint;
     public Vector3 spawnPos;
     public Quaternion spawnRot;
